// src/cdf-datasource/types.ts
export enum Tab {
  Timeseries = 'Timeseries',
  Asset = 'Asset',
  Custom = 'Custom',
  Event = 'Event',
  Relationships = 'Relationships',
  Templates = 'Templates',
  ExtractionPipelines = 'ExtractionPipelines',
  D3 = '3D Models',
}
// ...
export interface RelationshipsQuery {
  dataSetIds?: {
    id: number;
    value?: string;
  }[];
  labels?: {
    containsAny: {
      externalId: string;
      value?: string;
    }[];
  };
  isActiveAtTime?: boolean;
  activeAtTime?: {
    max: number;
    min: number;
  };
  sourceExternalIds?: string[];
  targetTypes?: string[];
  limit: number;
}
// ...
export interface ExtractionPipelinesQuery {
  ids?: {
    id?: number;
    value?: string;
  }[];
  numeric?: boolean;
  enableAllExtractionPipelines?: boolean;
}
// ...
export interface AssetQuery {
  target: string;
  includeSubtrees: boolean;
  withRelationships?: boolean;
  includeSubTimeseries?: boolean;
  relationshipsQuery?: RelationshipsQuery;
}

export interface EventQuery {
  expr: string;
  activeAtTimeRange: boolean;
  columns: string[];
  eventQuery: string;
}

export interface D3ModelQuery {
  modelId?: number | string;
  enableD3Models: boolean;
  project?: string;
}
export type CogniteQuery = CogniteQueryBase & CogniteTargetObj;

export interface CogniteQueryBase extends DataQuery {
  aggregation: string;
  granularity: string;
  latestValue: boolean;
  error: string;
  label: string;
  tab: Tab;
  assetQuery: AssetQuery;
  eventQuery: EventQuery;
  templateQuery: TemplateQuery;
  expr: string;
  warning: string;
  relationshipsQuery: RelationshipsQuery;
  extractionPipelinesQuery: ExtractionPipelinesQuery;
  d3ModelQuery: D3ModelQuery;
}

export type TemplateQuery = {
  groupExternalId: string;
  version: number;
  graphQlQuery: string;
  groupBy: string;
  datapointsPath: string;
  dataPath: string;
};

export type CogniteTargetObj =
  | {
      target?: number;
      targetRefType?: 'id';
    }
  | {
      target?: string;
      targetRefType?: 'externalId';
    };

export type QueryTarget = CogniteQuery;

export type QueryOptions = DataQueryRequest<CogniteQuery>;

export type Tuple<T> = [T, T];

/**
 * Comes from grafana, could be imported in future releases hopefully
 */
export interface AppEvent<T> {
  readonly name: string;
  payload?: T;
}

export enum HttpMethod {
  POST = 'POST',
  GET = 'GET',
  PATCH = 'PATCH',
  DELETE = 'DELETE',
}